[BUGFIX] Do not try build a request
Because of Maps2 Overlay we have our own
RequestHandler. It checks, if map should be shown
or not. In case of other plugins on same site
it may happen that maps2 will try to start the
RequestHandler of the foreign extension
which may result in an error. We have changed
that and get extKey of ConfigurationManager
directly which is more safe.